# Conflicts:
#	src/Union.Backend.API/Controllers/FriendsController.cs
#	src/Union.Backend.API/Controllers/GardensController.cs
#	src/Union.Backend.Service/Services/FriendsService.cs
#	src/Union.Backend.Service/Services/GardensService.cs

+++ src/Union.Backend.API/Controllers/GardenScoresController.cs
@@ -1,8 +1,11 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Union.Backend.Service.Services;
-using System;
+using Union.Backend.Service.Exceptions;
 using Union.Backend.Service.Dtos;
+using Union.Backend.Service.Auth;
+using System.Net;
 
 namespace Union.Backend.API.Controllers
 {
@@ -11,6 +14,7 @@ namespace Union.Backend.API.Controllers
     public class GardenScoresController : ControllerBase
     {
         private readonly ScoresService service;
+        private readonly GardensService GardenService;
         public GardenScoresController(ScoresService service)
         {
             this.service = service;
@@ -23,10 +27,31 @@ namespace Union.Backend.API.Controllers
         }
 
 
-        [HttpPost("{id}/score")] //TODO
+        [HttpPost("{id}/score")]
         public async Task<IActionResult> AddScore([FromRoute(Name = "id")] Guid GardenId, [FromBody] ScoreDto Score)
         {
-            return Created("TODO", await service.AddScore(Score, GardenId));
+            try
+            {
+                var id = Utils.ExtractIdFromToken(Request.Headers[HttpRequestHeader.Authorization.ToString()]);
+                var garden = await GardenService.GetGardenById(GardenId);
+                /* modifier dto
+                 if (garden.Data.Owner != id &&  garden.Data.Owner != id &&  !Utils.IsAdminRoleFromToken(Request.Headers[HttpRequestHeader.Authorization.ToString()]))
+                 {
+                     return Forbid();
+                 }*/
+                return Created("TODO", await service.AddScore(Score, GardenId));
+
+            }
+            catch (HttpResponseException)
+            {
+                throw;
+            }
+            catch (Exception)
+            {
+                throw new BadRequestApiException();
+            }
+
+
         }
 
         [HttpPost("score/{id}/report")] //TODO
@@ -38,7 +63,28 @@ namespace Union.Backend.API.Controllers
         [HttpDelete("score/{id}")]
         public async Task DeleteScore([FromRoute(Name = "id")] Guid ScoreId)
         {
-            await service.DeleteScore(ScoreId);
+
+            try
+            {
+                var id = Utils.ExtractIdFromToken(Request.Headers[HttpRequestHeader.Authorization.ToString()]);
+                var score = await service.GetScore(ScoreId);
+                /* modifier dto
+                if (score.Data.Rater == id || Utils.IsAdminRoleFromToken(Request.Headers[HttpRequestHeader.Authorization.ToString()]))
+                {
+                    await service.DeleteScore(ScoreId);
+                }*/
+
+            }
+            catch (HttpResponseException)
+            {
+                throw;
+            }
+            catch (Exception)
+            {
+                throw new BadRequestApiException();
+            }
+
+
         }
 
     }
